pyproject.toml
[project]
name = "pea-scanner"
version = "0.1.0"
description = "Bot de veille boursiere PEA - analyse technique, ML et alertes Telegram"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "loguru>=0.7.3",
    "pandas>=3.0.1",
    "pdfplumber>=0.11.9",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "yfinance>=0.2.50",
    "gnews>=0.4.3",
    "ta>=0.11.0",
    "xgboost>=2.1.0",
    "scikit-learn>=1.5.0",
    "joblib>=1.4.0",
    "numpy>=2.0.0",
    "openai>=1.0.0",
    "apscheduler>=3.10.0",
    "python-telegram-bot>=21.0",
    "feedparser>=6.0.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
